pyproject.toml
[tool.poetry]
name = "rewards"
version = "0.1.0"
description = ""
authors = ["derijablya <dulugov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
celery = "^5.5.1"
redis = "^5.2.1"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.7.0"
python-dotenv = "^1.1.0"
black = "^25.1.0"
uvicorn = "^0.34.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
